byline_parser: a parser working on top of the language
This parser is able to detect and extract comment and code parts of any
program, even the most awfuls one (see dedicated tests),
in order to simplify the work of the current parser that works on ASP
source code to extract the critical information itself.

This new parser, when ran before the precise/previous one, will be able
to simplify the parsing job and manipulate the full code.

current implementation uses a non-already existing feature of arpeggio
(multiline support).
However, as soon as arpeggio provides the modification, this code will
be usable by anyone.