major improvement on propagator system and behavior when clingo module available
various bugfixes + correct behaviors + complex propagator alchemy fixed.
A logical but disturbing behavior has been understood:
when using the clingo module to solve a file, any main defined in that
file will be ignored. Impossible to get a propagator defined in the file
to work while using the clingo module to embed ASP in python.